# grocy.openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Grocy REST API",
		"description": "Authentication is done via API keys (header *VICTUAL-API-KEY* or same named query parameter), which you can manage [here](PlaceHolderManageApiKeysUrl).<br>Additionally requests from within the frontend are also valid (via session cookie).",
		"version": "xxx",
		"license": {
			"name": "grocy.info",
			"url": "https://grocy.info"
		}
	},
	"servers": [
		{
			"url": "xxx"
		}
	],
...
	"paths": {
...
		"/stock/products/{productId}/open": {
			"post": {
				"summary": "Marks the given amount of the given product as opened",
				"tags": [
					"Stock"
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"description": "A valid product id",
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"amount": {
										"type": "number",
										"description": "The amount to mark as opened"
									},
									"stock_entry_id": {
										"type": "string",
										"description": "A specific stock entry id to open, if used, the amount has to be 1"
									},
									"allow_subproduct_substitution": {
										"type": "boolean",
										"description": "`true` when any in stock sub product should be used when the given product is a parent product and currently not in stock"
									}
								},
								"example": {
									"amount": 1
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "The operation was successful",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StockLogEntry"
									}
								}
							}
						}
					},
					"400": {
						"description": "The operation was not successful (possible errors are: Not existing product, given amount > current unopened stock amount)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error400"
								}
							}
						}
					}
				}
			}
		},
...
		"/stock/products/by-barcode/{barcode}/open": {
			"post": {
				"summary": "Marks the given amount of the by its barcode given product as opened",
				"tags": [
					"Stock \"by-barcode\""
				],
				"parameters": [
					{
						"in": "path",
						"name": "barcode",
						"required": true,
						"description": "Barcode",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"amount": {
										"type": "number",
										"description": "The amount to mark as opened"
									},
									"stock_entry_id": {
										"type": "string",
										"description": "A specific stock entry id to open, if used, the amount has to be 1"
									},
									"allow_subproduct_substitution": {
										"type": "boolean",
										"description": "`rue` when any in stock sub product should be used when the given product is a parent product and currently not in stock"
									}
								},
								"example": {
									"amount": 1
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "The operation was successful",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StockLogEntry"
									}
								}
							}
						}
					},
					"400": {
						"description": "The operation was not successful (possible errors are: Not existing product, given amount > current unopened stock amount)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error400"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"StockLogEntry": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"product_id": {
						"type": "integer"
					},
					"amount": {
						"type": "number"
					},
					"best_before_date": {
						"type": "string",
						"format": "date"
					},
					"purchased_date": {
						"type": "string",
						"format": "date"
					},
					"used_date": {
						"type": "string",
						"format": "date"
					},
					"spoiled": {
						"type": "boolean",
						"default": false
					},
					"stock_id": {
						"type": "string"
					},
					"transaction_id": {
						"type": "string"
					},
					"transaction_type": {
						"$ref": "#/components/schemas/StockTransactionType"
					},
					"note": {
						"type": "string"
					},
					"row_created_timestamp": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"Error400": {
				"type": "object",
				"properties": {
					"error_message": {
						"type": "string"
					}
				},
				"example": {
					"error_message": "The error message..."
				}
			},
...
			"StockTransactionType": {
				"type": "string",
				"enum": [
					"purchase",
					"consume",
					"inventory-correction",
					"product-opened"
				]
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "VICTUAL-API-KEY"
			}
		},
...
	},
	"security": [
		{
			"ApiKeyAuth": []
		}
	]
}